api/requirements.txt

Flask==1.0.2
uvloop==0.11.2
Flask-Cors==3.0.4
flask_restplus==0.12.1
flask_classy==0.6.7
flask-language==0.1.0
# Flask-MongoAlchemy==0.7.2
pymongo==3.5.1
Werkzeug==0.14.1
PyJWT==1.4.0